requirements.txt
torch>=2.0.0
watchdog
openai-whisper
